Move the source code of the decoder's public API to a clear location
