+++ src/decoder/complex.rs
@@ -5,10 +5,10 @@ use num::ToPrimitive;
 use self::missing::DecodedValue::{self, Missing1, Missing2, Normal};
 use crate::{
     decoder::{
-        common::*,
         param::{ComplexPackingParam, SimplePackingParam},
         simple::*,
         stream::{BitStream, NBitwiseIterator},
+        DecodeError, Grib2SubmessageDecoder,
     },
     error::*,
     utils::{read_as, GribInt},
