+++ src/decoder/jpeg2000.rs
@@ -3,7 +3,10 @@ use std::convert::TryInto;
 use openjpeg_sys as opj;
 
 use crate::{
-    decoder::{common::*, param::SimplePackingParam, simple::*, stream::FixedValueIterator},
+    decoder::{
+        param::SimplePackingParam, simple::*, stream::FixedValueIterator, DecodeError,
+        Grib2SubmessageDecoder,
+    },
     error::*,
     utils::read_as,
 };
